Fixes based on pull request comments
