+++ app.js
@@ -6,8 +6,8 @@ import express from 'express';
 import logger from 'morgan';
 import session from 'express-session';
 
-import utilsHelper from './helpers/utils';
-import getAccessTokenHelper from './helpers/accessToken';
+import { getAuthorizationUrl, getLogoutUrl} from './helpers/utils';
+import { getAccessToken } from './controllers/accessToken';
 
 const app = express();
 
@@ -21,11 +21,13 @@ const sessionConfig = {
   saveUninitialized: true,
   resave: true,
 };
+
 // session config for production
 if (app.get('env') === 'production') {
   app.set('trust proxy', 1); // trust first proxy
   sess.cookie.secure = true; // serve secure cookies
 }
+
 if (process.env.NODE_ENV !== 'test') {
   app.use(logger('dev'));
 }
@@ -39,27 +41,27 @@ app.get('/', (req, res) => {
   res.render('pages/index');
 });
 app.get('/login', (req, res) => {
-  res.redirect(utilsHelper.getAuthorizationUrl());
+  res.redirect(getAuthorizationUrl());
 });
 app.get('/callback', (req, res) => {
   // check if the mandatory Authorization code is there.
   if (!req.query.code) {
     res.sendStatus(400);
   }
-  getAccessTokenHelper.getAccessToken(res, req);
+  getAccessToken(res, req);
 });
 app.get('/profile', (req, res) => {
   const user = req.session.userInfo;
   res.render('pages/profile', { user });
 });
 app.get('/logout', (req, res) => {
-  res.redirect(utilsHelper.getLogoutUrl(req));
+  res.redirect(getLogoutUrl(req));
 });
 app.get('/end', (req, res) => {
   // resetting the id token hint.
   req.session.id_token = null;
   req.session.userInfo = null;
-  res.render('pages/end');
+  res.render('pages/logged-out');
 });
 
 // Starting server
