+++ config/config.json
@@ -1,9 +1,9 @@
 {
   "FC_URL": "https://fcp.integ01.dev-franceconnect.fr",
-  "AUTHORIZATION_URL": "https://fcp.dev.dev-franceconnect.fr/api/v1/authorize",
-  "TOKEN_URL": "https://fcp.dev.dev-franceconnect.fr/api/v1/token",
-  "USERINFO_URL": "https://fcp.dev.dev-franceconnect.fr/api/v1/userinfo",
-  "LOGOUT_URL": "https://fcp.dev.dev-franceconnect.fr/api/v1/logout",
+  "AUTHORIZATION_URL": "https://fcp.integ01.dev-franceconnect.fr/api/v1/authorize",
+  "TOKEN_URL": "https://fcp.integ01.dev-franceconnect.fr/api/v1/token",
+  "USERINFO_URL": "https://fcp.integ01.dev-franceconnect.fr/api/v1/userinfo",
+  "LOGOUT_URL": "https://fcp.integ01.dev-franceconnect.fr/api/v1/logout",
   "CLIENT_SECRET": "c48ff5ae96e870f507507555f7bc4dd361d2aac31df219fe6e92bbcca65f73f5",
   "SECRET_KEY": "8f373c6e6a48ce0f5931f414b6739e4e0aa82eda20a083dc5c0522b6c691b17b",
   "REDIRECT_URL": "http://localhost:3000/callback",
