+++ helpers/utils.js
@@ -5,7 +5,7 @@
  */
 import config from '../config/config.json';
 
-exports.getAuthorizationUrl = () => `${config.AUTHORIZATION_URL}?response_type=code`
+export const getAuthorizationUrl = () => `${config.AUTHORIZATION_URL}?response_type=code`
     + `&client_id=${config.CLIENT_SECRET}&redirect_uri=${config.REDIRECT_URL}`
     + `&scope=${config.SCOPE}&state=${config.STATE}&nonce=${config.NONCE}`;
 
@@ -14,5 +14,5 @@ exports.getAuthorizationUrl = () => `${config.AUTHORIZATION_URL}?response_type=c
  * Format the url 's that is used in a redirect call to France Connect logout API endpoint
  * @returns {string}
  */
-exports.getLogoutUrl = req => `${config.LOGOUT_URL}?id_token_hint=${req.session.id_token}`
+export const getLogoutUrl = req => `${config.LOGOUT_URL}?id_token_hint=${req.session.id_token}`
   + `&state=${config.STATE}&post_logout_redirect_uri=${config.LOGOUT_REDIRECT_URL}`;
