+++ views/pages/logged-out.ejs
@@ -6,7 +6,7 @@
                 Guichet des Chiroquois
             </h1>
             <div class="subtitle">
-                <h4>Merci de votre visite, veuillez fermer votre navigateur afin que la deconnection soit complète.</h4>
+                <h4>Merci de votre visite.</h4>
             </div>
         </div>
     </div>
