+++ views/pages/profile.ejs
@@ -15,8 +15,12 @@
                                 </a>
                             </div>
                             <div class="container has-text-centered">
-                                <p> Votre connexion au Guichet Chiroquois, via France Connect, est un succès</p>
-                                <p> Utiliser le button en haut gauche (avec votre nom et prénom), afin accèder à votre historique de connexion et au bouton de deconnexion .</p>
+                                <p class="has-text-centered"> Votre connexion au Guichet Chiroquois, via France Connect, est un succès</p>
+                                <p> Utiliser le button en haut gauche (avec votre nom et prénom), afin accèder à :</p>
+                                <div class="">
+                                    <p>Votre historique de connexion</p>
+                                    <p>Au bouton de deconnexion</p>
+                                </div>
                             </div>
                         </div>
                     </div>
